Create check_events rake task
Rake task which will replicate the behaviour initial implemented in the [gtm-utils repo](https://github.com/gclssvglx/gtm-utils).

The caller can specify an action of `test` or `create`. The 'test' code will diff the events fired with an expected event strucutre
and output (to the console) whether the fired-event structure is correct.

The 'create' code will output the events fired to a text file. This code is similar to the FakerJob code in this repo.
I chose to use 'create' for the naming as I felt like 'fake' implies that the events are not real or do not fire, which isn't the case.
In general in this repo, the name 'fake' is used because events are created without a user having to manually go to a URL and
trigger an event, they can do it from a central UI. I think this naming makes a bit more sense for the UI, but I think its confusing in the
context of outputting to a file, because real events are fired. The word 'fake' is used throughout the repo... I won't change it elsewhere
until i've had feedback on the naming, it might just be easier to keep everything as fake...

The 'create' and 'test' functionality will be implemented in future commits.

Some validation has been added to improve usability. This could be improved (for example checking that the iterations are a number with an upper limit).
For now i've made it optional and will do this by adding a default value in a future commit. It also breaks if there are any spaces, so that could be improved.

Example command: bundle exec rake check_events[create,integration,accordions]