Add TestEvents model
Code for testing the events. Methods that will also be used by the
upcoming CreateEvent code have been implemented in the parent class.

This code has been re-factored from its original implementation in
[this gtm-utils repo](https://github.com/gclssvglx/gtm-utils/blob/main/google_tag_manager.rb).

The implementation for tabs is currently broken and the URLs are not what they should be,
this makes the test fail. The implementation is in the process of being fixed which
may mean the test code needs tweaking.

There is not yet an implementation for comparing pageview events, only tabs and accordions.